requirements.txt
pynput
pyautogui
pyscreeze
pillow
pywinctl
pymonctl
numpy
matplotlib